+++ src/datafiles.py
@@ -3,6 +3,7 @@ import json
 import os
 
 
+# Check if a file exist, raise an error if not
 def check_file(filename):
 	if not os.path.exists(filename):
 		raise FileNotFoundError(f"File <{filename}> not found.")
@@ -15,7 +16,7 @@ def read_csv(filename):
 	# Read an CSV file and return a list of dictionaries with name_column: value
 	check_file(filename)
 	output = []
-	with open(filename) as f:
+	with open(filename, encoding="utf-8") as f:
 		data = csv.DictReader(f)
 		for row in data:
 			output.append(row)
@@ -28,14 +29,9 @@ def read_csv(filename):
 def read_json(filename):
 	# Read a JSON file and return a dictionary with the content
 	check_file(filename)
-	with open(filename, 'r') as f:
+	with open(filename, encoding="utf-8") as f:
 		data = json.load(f)
 	return data
-
-
-def read_json_key(filename, key):
-	# Read a JSON file ando return the value of a given key
-	return read_json(filename)[key]
 
 
 # TXT
@@ -44,7 +40,7 @@ def read_json_key(filename, key):
 def read_txt(filename):
 	# Read the content of a text file
 	check_file(filename)
-	with open(filename) as f:
+	with open(filename, encoding="utf-8") as f:
 		output = f.read()
 	return output