+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 1.1 (2022-05-05)
+
+- Added support for UTF-8 charset
+- Added support for HTML in mail content files
+
+
 ## Version 1.0 (2022-05-04)
 
 - Take files as parameters.
@@ -8,7 +14,7 @@
 ## Version 0.1a (2022-05-03)
 
 - Read mail settings, recipients list and mail content from files.
-- Send the mails.
+- Sends the mails.
 
 
 ## (2022-04-27)